Dynamodb_migrations/dynamo_migrator/requirements.txt
boto3
tqdm
pyyaml
